requirements.txt
Django >= 1.5, <1.8
djangorestframework